fix some issues mentioned in the comment

+++ src/main/java/com/example/spring/board/services/BookingService.java
@@ -4,7 +4,9 @@ import com.example.spring.board.dto.req.CreateBooking;
 import com.example.spring.board.dto.res.BookingDetail;
 import com.example.spring.board.model.Booking;
 import com.example.spring.board.services.core.BookingCoreService;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -31,33 +33,32 @@ public class BookingService {
         List<Booking> bookings = bookingCoreService.getAllBooking();
         List<BookingDetail> bookingDetails = new ArrayList<>();
 
-        for(Booking b : bookings){
-            BookingDetail bookingDetail = new BookingDetail();
-
-            bookingDetail.setId(b.getId());
-            bookingDetail.setStartDate(b.getStartDate());
-            bookingDetail.setEndDate(b.getEndDate());
-
+        bookings.forEach(b -> {
+            BookingDetail bookingDetail = new BookingDetail(
+                    b.getId(),
+                    b.getStartDate(),
+                    b.getEndDate()
+            );
             bookingDetails.add(bookingDetail);
-        }
-
+        });
         return ResponseEntity.ok(bookingDetails);
     }
 
     public BookingDetail getBookingByIdService(Long id){
         Booking booking = bookingCoreService.getBookingById(id);
-        BookingDetail bookingDetail = new BookingDetail();
-
-        bookingDetail.setId(booking.getId());
-        bookingDetail.setStartDate(booking.getStartDate());
-        bookingDetail.setEndDate(booking.getEndDate());
-        System.out.println(booking.getEndDate());
-
-        return bookingDetail;
+        return new BookingDetail(
+                booking.getId(),
+                booking.getStartDate(),
+                booking.getEndDate()
+        );
     }
 
-    public void deleteBookingByIdService(Long id){
+    public ResponseEntity<Booking> deleteBookingByIdService(Long id){
+        Booking booking = bookingCoreService.getBookingById(id);
+        if(booking == null){
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).build(); // Return 404 if booking doesn't exist
+        }
         bookingCoreService.deleteBookingById(id);
+        return ResponseEntity.ok(booking);
     }
-
 }
