+++ src/main/java/com/example/spring/board/services/VehicleTypeService.java
@@ -3,13 +3,16 @@ package com.example.spring.board.services;
 import com.example.spring.board.dto.res.VehicleTypeDetail;
 import com.example.spring.board.model.VehicleType;
 import com.example.spring.board.services.core.VehicleTypeCoreService;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
@@ -32,8 +35,14 @@ public class VehicleTypeService {
     }
 
 
-    public void deleteVehicleTypeByIdService(Long id){
+    public ResponseEntity<VehicleType> deleteVehicleTypeByIdService(Long id){
+        VehicleType vehicleType = vehicleTypeCoreService.getVehicleTypeById(id);
+        if(Objects.isNull(vehicleType)){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+
         vehicleTypeCoreService.deleteVehicleTypeById(id);
+        return ResponseEntity.ok(vehicleType);
     }
 
     public ResponseEntity<List<VehicleTypeDetail>> getAllVehicleTypeService(){
